Fix CI typing and align site nav and theme toggle
- Resolve mypy errors (float return, Mapping narrowing, baseline prober
  protocol signature, ingest CLI types) that broke the CI run.
- Register launch docs in the public-surface allowlist.
- Point board/solved Data and Docs nav at the bare repos (no tree/main).
- Add the day/night theme toggle to the generated board and solved pages
  so they match the homepage, sharing the lemma-theme preference key.

+++ lemma/scoring/rolling.py
@@ -37,7 +37,7 @@ def decay_weight(age_epochs: int, *, half_life_epochs: float) -> float:
     """Return the exponential decay multiplier for credit ``age_epochs`` old."""
     if age_epochs < 0:
         raise ValueError("age_epochs must be non-negative")
-    return 0.5 ** (age_epochs / half_life_epochs)
+    return float(0.5 ** (age_epochs / half_life_epochs))
 
 
 def rolling_scores(
